+++ twitter4j-core/src/main/java/twitter4j/IDs.java
@@ -23,7 +23,7 @@ package twitter4j;
  * @author Yusuke Yamamoto - yusuke at mac.com
  */
 public interface IDs extends TwitterResponse, CursorSupport, java.io.Serializable {
-    int[] getIDs();
+    long[] getIDs();
 
     boolean hasPrevious();
